# Conflicts:
#	Content.Shared/CCVar/CCVars.cs
#	Resources/Prototypes/Entities/Mobs/Species/human.yml

+++ Content.Server/Construction/Completions/DeleteEntity.cs
@@ -1,18 +1,29 @@
-﻿using System.Threading.Tasks;
 using Content.Shared.Construction;
 using JetBrains.Annotations;
-using Robust.Shared.GameObjects;
-using Robust.Shared.Serialization.Manager.Attributes;
 
 namespace Content.Server.Construction.Completions
 {
+    public sealed class ConstructionBeforeDeleteEvent : CancellableEntityEventArgs
+    {
+        public EntityUid? User;
+
+        public ConstructionBeforeDeleteEvent(EntityUid? user)
+        {
+            User = user;
+        }
+    }
+
     [UsedImplicitly]
     [DataDefinition]
-    public class DeleteEntity : IGraphAction
+    public sealed class DeleteEntity : IGraphAction
     {
         public void PerformAction(EntityUid uid, EntityUid? userUid, IEntityManager entityManager)
         {
-            entityManager.DeleteEntity(uid);
+            var ev = new ConstructionBeforeDeleteEvent(userUid);
+            entityManager.EventBus.RaiseLocalEvent(uid, ev);
+
+            if (!ev.Cancelled)
+                entityManager.DeleteEntity(uid);
         }
     }
 }
